# Speed up GetImplementationTypes