PR tries to speedup the `ImplementationTypesCollection.GetImplementationTypes`.

It uses `Type.IsDefined` to check if a type is ignorable implementation or not.

Here is the benchmark which checks N(in our case N is 25090) types for `Type.IsDefined`:
```
using System.Runtime.Serialization;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

namespace Attributes
{
    public class EntyPoint
    {
        public static void Main()
        {
            BenchmarkRunner.Run<Attributes>();
        }
    }

    [DataContract]
    public class A
    {
    }

    [ClrJob]
    public class Attributes
    {
        [Benchmark]
        public int IsDefined()
        {
            var count = 0;
            for (var i = 0; i < 25090; ++i) count += typeof(A).IsDefined(typeof(DataContractAttribute), false) ? 1 : 0;
            return count;
        }
    }
}
```

Here is the result:
```
BenchmarkDotNet=v0.11.2, OS=Windows 10.0.17134.345 (1803/April2018Update/Redstone4)
Intel Xeon Processor (Skylake, IBRS), 1 CPU, 2 logical and 2 physical cores
  [Host] : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3190.0
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3190.0

Job=Clr  Runtime=Clr

    Method |     Mean |     Error |   StdDev |
---------- |---------:|----------:|---------:|
 IsDefined | 39.66 ms | 0.9180 ms | 2.678 ms |
```

For every first resolve of the abstraction(it's actually every interface and in our case we have a lot of interfaces) needs at least 40ms only for `Type.IsDefined` calls.

PS Also `Type.IsDefined` produces a lot of allocations:
![image](https://user-images.githubusercontent.com/664889/48479876-a5e6e880-e819-11e8-8fe5-16d43d3e1a72.png)